Refactor REPL IDE services testing configuration and add new tests
